fix: correct the glyph cache hash, and clear the remaining diagnostics
std::hash<text::glyph_key> applied its masks after the shifts rather than
before:

    ((size_t(f) << 32) & 0xFFFF) | ((i << 16) & 0xFFFF) | (s & 0xFF)

0xFFFF covers bits 0..15, so both shifted components were masked away entirely
and every glyph key hashed to the low 8 bits of the point size alone. Font face
and glyph index contributed nothing, so with a single font size in use the
rasterizer's glyph cache collapsed to one bucket and every lookup degenerated to
a linear scan of the whole cache. Each component is now masked before being
shifted into a disjoint field. Equality is decided by operator==, so this is a
distribution fix, not a correctness one.

The rest of this commit clears the last ~300 clang-tidy reports across every
module: thread-safe replacements for strerror/getpwuid/localtime/rand, overrides
moved into the base's access section, const& parameters, widening applied to the
operands rather than the result, scoped enums, and named-enumerator defaults in
place of {} on enums without a zero value.

Three fixes needed correcting on top of the automated output:

- crispy/algorithm.h: std::ranges::for_each rejects a forwarded rvalue that is
  not a borrowed_range. The range is bound to a named reference instead, which
  also fixes the use-after-move the check was reporting.
- contour/Config.h: constifying the variadic process() parameters meant they no
  longer bound to the format() overloads taking a non-const reference, so
  ArcStyle, BrailleStyle and GitDrawingsStyle fell through to the generic
  std::format path and failed to compile. Those overloads now take const&.
- text_shaper/font.h: see above.

+++ src/vtbackend/regis/ReGISRasterizer.cpp
@@ -320,7 +320,8 @@ void ReGISRasterizer::blendCoverage(Pen const& pen,
     for (auto ty = 0; ty < h; ++ty)
         for (auto tx = 0; tx < w; ++tx)
         {
-            auto const cov = coverage[static_cast<size_t>((ty * w) + tx)];
+            auto const cov =
+                coverage[(static_cast<size_t>(ty) * static_cast<size_t>(w)) + static_cast<size_t>(tx)];
             if (cov != 0)
                 compositePixel(origin.x + tx, origin.y + ty, pen.color, pen.mode, cov);
         }
