+++ src/vtrasterizer/utils.cpp
@@ -24,7 +24,7 @@ vector<uint8_t> downsampleRGBA(vector<uint8_t> const& bitmap,
     auto const factor = static_cast<unsigned>(ceil(ratio));
 
     vector<uint8_t> dest;
-    dest.resize(static_cast<std::size_t>(*newSize.height * *newSize.width * 4));
+    dest.resize(static_cast<std::size_t>(*newSize.height) * static_cast<std::size_t>(*newSize.width) * 4);
 
     // RasterizerLog()("scaling from {} to {}, ratio {}x{} ({}), factor {}",
     //                 size, newSize, ratioX, ratioY, ratio, factor);
@@ -43,8 +43,8 @@ vector<uint8_t> downsampleRGBA(vector<uint8_t> const& bitmap,
             unsigned int count = 0;
             for (unsigned y = sr; y < min(sr + factor, size.height.as<unsigned>()); y++)
             {
-                uint8_t const* p =
-                    bitmap.data() + (y * unbox(size.width) * 4) + (static_cast<size_t>(sc * 4));
+                uint8_t const* p = bitmap.data() + (static_cast<size_t>(y) * unbox(size.width) * 4)
+                                   + (static_cast<size_t>(sc) * 4);
                 for (unsigned x = sc; x < min(sc + factor, size.width.as<unsigned>()); x++, count++)
                 {
                     b += *(p++);
@@ -102,8 +102,8 @@ vector<uint8_t> downsample(vector<uint8_t> const& bitmap,
             unsigned count = 0; // number of pixels being averaged
             for (auto y = sr; y < min(sr + factor, size.height.as<unsigned>()); y++)
             {
-                uint8_t const* p = bitmap.data() + (y * *size.width * numComponents)
-                                   + (static_cast<size_t>(sc * numComponents));
+                uint8_t const* p = bitmap.data() + (static_cast<size_t>(y) * *size.width * numComponents)
+                                   + (static_cast<size_t>(sc) * numComponents);
                 for (auto x = sc; x < min(sc + factor, size.width.as<unsigned>()); x++, count++)
                     for (auto const k: std::views::iota(0u, numComponents))
                         values.at(k) += *(p++);
